PageWithReviews/requirements.txt
django >= 2.1.6, < 3.0
psycopg2 == 2.8.3
djangorestframework == 3.10.2